GmailClasiffer/app/requirements.txt
google-genai==0.3.0     
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.141.0
google-cloud-aiplatform==1.70.0
google-cloud-secret-manager==2.20.2